refactor: isolate Modal stories
