+++ packages/react-aria-components/stories/Modal.stories.tsx
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2022 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import {Button, Dialog, DialogTrigger, Heading, Modal, ModalOverlay} from 'react-aria-components';
+import React from 'react';
+
+export default {
+  title: 'React Aria Components'
+};
+
+export const ModalExample = () => (
+  <DialogTrigger>
+    <Button>Open modal</Button>
+    <ModalOverlay
+      style={{
+        position: 'fixed',
+        zIndex: 100,
+        top: 0,
+        left: 0,
+        bottom: 0,
+        right: 0,
+        background: 'rgba(0, 0, 0, 0.5)',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center'
+      }}>
+      <Modal
+        style={{
+          background: 'Canvas',
+          color: 'CanvasText',
+          border: '1px solid gray',
+          padding: 30
+        }}>
+        <Dialog>
+          {({close}) => (
+            <form style={{display: 'flex', flexDirection: 'column'}}>
+              <Heading slot="title" style={{marginTop: 0}}>Sign up</Heading>
+              <label>
+                First Name: <input placeholder="John" />
+              </label>
+              <label>
+                Last Name: <input placeholder="Smith" />
+              </label>
+              <Button onPress={close} style={{marginTop: 10}}>
+                Submit
+              </Button>
+            </form>
+          )}
+        </Dialog>
+      </Modal>
+    </ModalOverlay>
+  </DialogTrigger>
+);
