+++ packages/react-aria-components/stories/index.stories.tsx
@@ -472,50 +472,6 @@ export const PopoverExample = () => (
   </DialogTrigger>
 );
 
-export const ModalExample = () => (
-  <DialogTrigger>
-    <Button>Open modal</Button>
-    <ModalOverlay
-      style={{
-        position: 'fixed',
-        zIndex: 100,
-        top: 0,
-        left: 0,
-        bottom: 0,
-        right: 0,
-        background: 'rgba(0, 0, 0, 0.5)',
-        display: 'flex',
-        alignItems: 'center',
-        justifyContent: 'center'
-      }}>
-      <Modal
-        style={{
-          background: 'Canvas',
-          color: 'CanvasText',
-          border: '1px solid gray',
-          padding: 30
-        }}>
-        <Dialog>
-          {({close}) => (
-            <form style={{display: 'flex', flexDirection: 'column'}}>
-              <Heading slot="title" style={{marginTop: 0}}>Sign up</Heading>
-              <label>
-                First Name: <input placeholder="John" />
-              </label>
-              <label>
-                Last Name: <input placeholder="Smith" />
-              </label>
-              <Button onPress={close} style={{marginTop: 10}}>
-                Submit
-              </Button>
-            </form>
-          )}
-        </Dialog>
-      </Modal>
-    </ModalOverlay>
-  </DialogTrigger>
-);
-
 export const TabsExample = () => {
   let [url, setUrl] = useState('/FoR');
